refactor: move ToolBase to mcp-core, atlas-local tools to mcp-tools-atlas-local MCP-480
- Move ToolBase to @mongodb-js/mcp-core, decoupled from concrete types:
  - Session → IToolSession interface
  - UserConfig → IToolConfig interface
  - Telemetry → ITelemetry interface
  - Elicitation → IElicitation interface
  - Server → inline { mcpServer: McpServer }
- Add IToolConfig, IToolSession, PreviewFeature, TransportType to @mongodb-js/mcp-types
- Add telemetry event types (ConnectionMetadata, TelemetryToolMetadata, ToolEvent,
  AtlasMetadata, etc.) to @mongodb-js/mcp-types
- Add getRandomUUID, TRANSPORT_PAYLOAD_LIMITS, CommonArgs, formatUntrustedData to mcp-core
- Move AtlasLocalToolBase + 4 tools to @mongodb-js/mcp-tools-atlas-local
  - Define IAtlasLocalConfig and IAtlasLocalSession in the package
  - Export AtlasLocalTools: ToolClass[]
- Delete src/common/atlasLocal.ts and src/tools/atlasLocal/ (no re-export shims)
- Update all callers to import directly from @mongodb-js/mcp-tools-atlas-local
- src/tools/tool.ts is now a re-export shim from @mongodb-js/mcp-core
