+++ packages/core/package.json
@@ -26,12 +26,17 @@
     "test": "vitest --run"
   },
   "dependencies": {
+    "@mcp-ui/server": "^5.13.1",
     "@modelcontextprotocol/sdk": "^1.29.0",
+    "bson": "^7.2.0",
+    "@mongodb-js/mcp-logging": "workspace:*",
+    "@mongodb-js/mcp-metrics": "workspace:*",
     "@mongodb-js/mcp-types": "workspace:*",
-    "mongodb-redact": "^1.1.6"
+    "@mongodb-js/mcp-ui": "workspace:*",
+    "mongodb-redact": "^1.1.6",
+    "zod": "^4.0.0"
   },
   "devDependencies": {
-    "@mongodb-js/mcp-logging": "workspace:*",
     "@types/node": "^25.6.0",
     "typescript": "^5.9.2",
     "vitest": "^4.1.4"
