+++ packages/core/src/index.ts
@@ -17,6 +17,23 @@ export { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 
 import type { LogLevel } from "@mongodb-js/mcp-types";
 
+export {
+    ToolBase,
+    type ToolClass,
+    type ToolConstructorParams,
+    type AnyToolBase,
+    type ToolArgs,
+    type ToolExecutionContext,
+    type ToolResult,
+    type OperationType,
+    type ToolCategory,
+    formatUntrustedData,
+} from "./tool.js";
+export type { DefaultMetrics } from "@mongodb-js/mcp-metrics";
+export { getRandomUUID } from "./randomUUID.js";
+export { TRANSPORT_PAYLOAD_LIMITS } from "./transportConstants.js";
+export { CommonArgs, NO_UNICODE_ERROR } from "./args.js";
+
 export const MCP_LOG_LEVELS: readonly LogLevel[] = [
     "debug",
     "info",
