+++ packages/core/tsconfig.json
@@ -9,5 +9,5 @@
   },
   "include": ["src/**/*.ts"],
   "exclude": ["dist", "node_modules", "src/**/*.test.ts"],
-  "references": [{ "path": "../types" }]
+  "references": [{ "path": "../types" }, { "path": "../metrics" }, { "path": "../ui" }]
 }
