+++ packages/metrics/tsconfig.json
@@ -4,7 +4,8 @@
     "rootDir": "./src",
     "outDir": "./dist",
     "baseUrl": ".",
-    "erasableSyntaxOnly": true
+    "erasableSyntaxOnly": true,
+    "composite": true
   },
   "include": ["src/**/*.ts"],
   "exclude": ["dist", "node_modules", "src/**/*.test.ts"]
