+++ packages/tools-atlas-local/src/atlasLocalTool.ts
@@ -1,14 +1,26 @@
 import type { CallToolResult } from "@modelcontextprotocol/sdk/types.js";
-import type { ToolArgs, ToolCategory, ToolExecutionContext } from "../tool.js";
-import { ToolBase } from "../tool.js";
+import type { ToolArgs, ToolExecutionContext } from "@mongodb-js/mcp-core";
+import { ToolBase } from "@mongodb-js/mcp-core";
 import type { Client } from "@mongodb-js/atlas-local";
 import { LogId } from "@mongodb-js/mcp-logging";
-import type { ConnectionMetadata } from "../../telemetry/types.js";
+import type { ConnectionMetadata, IToolConfig, IToolSession } from "@mongodb-js/mcp-types";
+import type { ToolCategory } from "@mongodb-js/mcp-types";
+
+export interface IAtlasLocalConfig extends IToolConfig {
+    voyageApiKey?: string;
+}
+
+export interface IAtlasLocalSession extends IToolSession {
+    readonly atlasLocalClient?: Client;
+    connectToMongoDB(settings: { connectionString: string }): Promise<void>;
+}
 
 export const AtlasLocalToolMetadataDeploymentIdKey = "deploymentId";
 
-export abstract class AtlasLocalToolBase extends ToolBase {
+export abstract class AtlasLocalToolBase extends ToolBase<IAtlasLocalConfig> {
     static category: ToolCategory = "atlas-local";
+
+    declare protected readonly session: IAtlasLocalSession;
 
     protected verifyAllowed(): boolean {
         return this.session.atlasLocalClient !== undefined && super.verifyAllowed();
