+++ packages/tools-atlas-local/src/tools/connect/connectDeployment.ts
@@ -1,9 +1,9 @@
 import type { CallToolResult } from "@modelcontextprotocol/sdk/types.js";
-import { AtlasLocalToolBase } from "../atlasLocalTool.js";
-import type { OperationType, ToolArgs } from "../../tool.js";
+import { AtlasLocalToolBase } from "../../atlasLocalTool.js";
+import type { OperationType, ToolArgs } from "@mongodb-js/mcp-core";
 import type { Client } from "@mongodb-js/atlas-local";
-import { CommonArgs } from "../../args.js";
-import type { ConnectionMetadata } from "../../../telemetry/types.js";
+import { CommonArgs } from "@mongodb-js/mcp-core";
+import type { ConnectionMetadata } from "@mongodb-js/mcp-types";
 
 export class ConnectDeploymentTool extends AtlasLocalToolBase {
     static toolName = "atlas-local-connect-deployment";
