+++ packages/tools-atlas-local/src/tools/tools.ts
@@ -1,4 +1,4 @@
-export { DeleteDeploymentTool } from "./delete/deleteDeployment.js";
-export { ListDeploymentsTool } from "./read/listDeployments.js";
 export { CreateDeploymentTool } from "./create/createDeployment.js";
+export { ListDeploymentsTool } from "./read/listDeployments.js";
+export { DeleteDeploymentTool } from "./delete/deleteDeployment.js";
 export { ConnectDeploymentTool } from "./connect/connectDeployment.js";
