+++ packages/types/src/index.ts
@@ -1,3 +1,5 @@
+export type * from "./config.js";
+export type * from "./schemas.js";
 export type * from "./session.js";
 export type * from "./sessionStore.js";
 export type * from "./keychain.js";
