+++ packages/types/src/schemas.ts
@@ -1,0 +1,1 @@
+export type PreviewFeature = "mcpUI";
