--- src/common/atlasLocal.ts
@@ -1,5 +1,0 @@
-export {
-    defaultCreateAtlasLocalClient,
-    type AtlasLocalClientFactoryFn,
-    type LibraryLoader,
-} from "@mongodb-js/mcp-tools-atlas-local";
